+++ assets/pages/new-contributions-landing/components/ContributionThankYou/ContributionThankYou.jsx
@@ -7,10 +7,10 @@ import type { PaymentMethod } from 'helpers/contributions';
 import React from 'react';
 import { connect } from 'react-redux';
 import { type ContributionType, getSpokenType } from 'helpers/contributions';
-import { ButtonWithRightArrow } from './ButtonWithRightArrow/ButtonWithRightArrow';
-import MarketingConsent from './MarketingConsentContainer';
-import ContributionsSurvey from './ContributionSurvey/ContributionsSurvey';
-import { type Action, setHasSeenDirectDebitThankYouCopy } from '../contributionsLandingActions';
+import { ButtonWithRightArrow } from '../ButtonWithRightArrow/ButtonWithRightArrow';
+import MarketingConsent from '../MarketingConsentContainer';
+import ContributionsSurvey from '../ContributionSurvey/ContributionsSurvey';
+import { type Action, setHasSeenDirectDebitThankYouCopy } from '../../contributionsLandingActions';
 
 // ----- Types ----- //
