+++ assets/pages/new-contributions-landing/components/ContributionThankYou/ContributionThankYouContainer.jsx
@@ -4,10 +4,10 @@
 
 import React from 'react';
 import { connect } from 'react-redux';
-import { type ThankYouPageStageMap, type ThankYouPageStage } from '../contributionsLandingReducer';
+import { type ThankYouPageStageMap, type ThankYouPageStage } from '../../contributionsLandingReducer';
 import ContributionThankYou from './ContributionThankYou';
 import ContributionThankYouSetPassword from './ContributionThankYouSetPassword';
-import ContributionThankYouPasswordSet from './ContributionThankYouPasswordSet/ContributionThankYouPasswordSet';
+import ContributionThankYouPasswordSet from './ContributionThankYouPasswordSet';
 
 // ----- Types ----- //
