+++ assets/pages/new-contributions-landing/components/ContributionThankYou/ContributionThankYouSetPassword.jsx
@@ -6,8 +6,8 @@ import { type Dispatch } from 'redux';
 import React from 'react';
 import type { PaymentMethod } from 'helpers/contributions';
 import { connect } from 'react-redux';
-import { type Action, setHasSeenDirectDebitThankYouCopy } from '../contributionsLandingActions';
-import SetPasswordForm from './SetPasswordForm';
+import { type Action, setHasSeenDirectDebitThankYouCopy } from '../../contributionsLandingActions';
+import SetPasswordForm from '../SetPasswordForm';
 
 // ----- Types ----- //
