--- assets/pages/new-contributions-landing/components/ContributionThankYouPasswordSet/ContributionThankYouPasswordSet.scss
@@ -1,7 +1,0 @@
-// -----  gu-sass ----- //
-
-.thank-you__container {
-  @include mq($from: mobile, $until: tablet) {
-    background-color: gu-colour(news-garnett-highlight);
-  }
-}
